Reorganize the code a little (#1)
* Light code refactoring
