+++ main.py
@@ -1,18 +1,16 @@
 #!/usr/bin/env python3.8
 import sys
 import os
+
+from converter import PassThruRpzConverter, converters
+from processor import RpzProcessor
+
 try:
     import pwd
 except ModuleNotFoundError:
     pass
-from abc import ABC, abstractmethod
 from argparse import ArgumentParser
-from time import time
-from typing import TextIO
 from urllib import parse
-
-import requests
-from requests import HTTPError
 
 """
 Retrieve a DNS blackhole list from a given URL, optionally apply an allow-list, and write the list out in
@@ -26,275 +24,36 @@ DEFAULT_ALLOW_LIST_FILE = '/usr/local/etc/namedb/rpz-allowlist'
 MAX_DOMAIN_LENGTH = 240
 
 
-class RpzConverter(ABC):
-    """
-    Abstract Base Class for RPZ writers so that different strategies may be employed based on the type of data
-    being read from the blocklist.
-    """
-
-    @abstractmethod
-    def before_writing(self, output: TextIO):
-        """
-        Called after the RPZ file has been opened for writing, but before any data has been written to the file.
-        """
-        pass
-
-    @abstractmethod
-    def write_line(self, output: TextIO, line: str):
-        """
-        Called for every line that needs to be written to the RPZ file.
-        """
-        pass
-
-    @abstractmethod
-    def after_writing(self, output: TextIO):
-        """
-        Called once all the lines have been written, but the file is still open.  (Do not close the file here;
-        the RpzProcessor will handle that.)
-        """
-        pass
-
-    @abstractmethod
-    def should_passthru(self, line: str):
-        """
-        Called once per line; return true if the line should be passed through, or false if it probably contains
-        a domain which should be evaluated
-        """
-
-    def extract_domain(self, line: str):
-        """
-        Given a line from the source file, extract just the domain from the line.
-        """
-        return line.split()[0]
-
-    @staticmethod
-    @abstractmethod
-    def get_name():
-        """
-        Name for this converter, suitable for consumption as a command-line argument.
-        """
-        pass
-
-
-class PassThruRpzConverter(RpzConverter):
-    """
-    Pass-through writer for files which are already in RPZ format.  Does nothing special.
-    """
-    def before_writing(self, output: TextIO):
-        pass
-
-    def after_writing(self, output: TextIO):
-        pass
-
-    def should_passthru(self, line: str):
-        # cheap & easy skip DNS directives and comments
-        return line[0] in (';', '$', '@', ' ')
-
-    def write_line(self, output: TextIO, line: str):
-        output.write(line)
-        output.write('\n')
-
-    @staticmethod
-    def get_name():
-        return 'rpz'
-
-
-class DomainConverter(RpzConverter):
-    """
-    Convert from a hash-commented, domain-per-line text file to an RPZ file.  Generates a barebones zone file
-    header using the current epoch time as the serial number.
-    """
-
-    PREAMBLE = f'''$TTL 2h
-@ IN SOA localhost. root.localhost. ({int(time())} 6h 1h 1w 2h)
-  IN NS  localhost.
-'''
-
-    def extract_domain(self, line: str):
-        return line
-
-    def before_writing(self, output: TextIO):
-        output.write(DomainConverter.PREAMBLE)
-        output.write('\n')
-
-    def after_writing(self, output: TextIO):
-        pass
-
-    def should_passthru(self, line: str):
-        return line[0] in (';', '#')
-
-    def write_line(self, output: TextIO, line: str):
-        # convert hashes to semicolons for RPZ
-        if line[0] == '#':  # Faster than line.startswith('#')
-            line = line.replace('#', ';', 1)  # Faster than ';' + line[1:]
-            output.write(line)
-            output.write('\n')
-        else:
-            output.write(line)
-            output.write(' CNAME .\n')
-
-    @staticmethod
-    def get_name():
-        return 'domains'
-
-
-class WildcardDomainConverter(DomainConverter):
-    """
-    Convert from a hash-commented, wildcard-domain-per-line text file to an RPZ file.
-    Specifically, this should be used for wildcard domain files that include the wildcard
-    domain, but not the domain itself.   For example, *.example.com, but not example.com.
-
-    Because BIND RPZ does not match `example.com` on `*.example.com`, it is necessary to
-    output both the wildcard line, and the bare domain line as well; ie, for `*.example.com`,
-    RRs for both `*.example.com` and `example.com` will be written.
-    """
-    def extract_domain(self, line: str):
-        return line[2:] if line[0:2] == '*.' else line
-
-    def write_line(self, output: TextIO, line: str):
-        # convert hashes to semicolons for RPZ
-        if line[0] == '#':  # Faster than line.startswith('#')
-            line = line.replace('#', ';', 1)  # Faster than ';' + line[1:]
-            output.write(line)
-            output.write('\n')
-        else:
-            output.write(line)
-            output.write(' CNAME .\n')
-            # For wildcard domains, also write out the bare domain for BIND.
-            if line[0:2] == '*.':
-                output.write(line[2:])
-                output.write(' CNAME .\n')
-
-    @staticmethod
-    def get_name():
-        return 'wildcards'
-
-
-class RpzProcessor:
-    """
-    Class for importing a Response Policy Zone (RPZ) file from a URL, optionally applying an Allow List to ignore
-    matching lines in that file, then writing the output to a local file.
-    """
-
-    allow_domains_exact = set()
-    allow_domains_right = set()
-    converter = None
-
-    def __init__(self, converter: RpzConverter):
-        self.converter = converter
-
-    def read_allow_list(self, allow_list_file: str):
-        """
-        Read an allow-list file.  The format of this file is as follows:
-
-        # Comments start at the beginning of a line with a hash symbol
-        #
-        # Exact match domains begin with any valid character other than a dot (.)
-        example.com
-        #
-        # Lines which begin with a dot (.) are "right-hand" matches and also exact matches:
-        .example.net
-        # The above will match foo.example.net, foo.bar.baz.example.net and example.net.
-        .badness.example.edu
-        # The above will match some.badness.example.edu and badness.example.edu, but not example.edu.
-
-        :param allow_list_file: the path to the file containing the Allow List.
-        :return: True if reading the file was successful; false otherwise.
-        """
-        try:
-            with open(allow_list_file, 'r') as allow_list:
-                while line := allow_list.readline():
-                    line = line.strip()
-                    if len(line) < 2 or line.startswith('#'):
-                        continue
-                    if line.startswith('.'):
-                        line = line[1:]
-                        self.allow_domains_right.add(line)
-                        self.allow_domains_exact.add(line)
-                    else:
-                        self.allow_domains_exact.add(line)
-            return True
-        except OSError as e:
-            print(f'Failed to read allow-list file {allow_list_file}: {e}')
-        return False
-
-    def import_rpz_list(self, rpz_url: str, output_file: str):
-        try:
-            self._do_import(rpz_url, output_file)
-            return True
-        except HTTPError as e:
-            print(f'Failed to retrieve {rpz_url}: {e}')
-        except OSError as e:
-            print(f'Unable to open or write {output_file}: {e}')
-
-        return False
-
-    def _do_import(self, rpz_url: str, output_file: str):
-        session = requests.Session()
-
-        with session.get(rpz_url, stream=True) as request:
-            request.raise_for_status()  # in case the response is not a 200
-
-            with open(output_file, 'w') as output:
-
-                self.converter.before_writing(output)
-
-                for line in request.iter_lines(chunk_size=256, decode_unicode=True):
-                    if not line:  # skip blank lines
-                        continue
-
-                    if self.converter.should_passthru(line):
-                        self.converter.write_line(output, line)
-                        continue
-
-                    domain = self.converter.extract_domain(line)
-
-                    # sometimes there's garbage in an RPZ file and the resulting domain name is > 255 characters long
-                    # including the base domain name (e.g., 'localhost').  This is a cheap attempt to ignore very long
-                    # domain names.
-                    if len(domain) > MAX_DOMAIN_LENGTH:
-                        continue
-
-                    if domain in self.allow_domains_exact:
-                        continue
-
-                    # check for a right-hand match by decomposing the domain into successive subdomains and checking each
-                    # against the right-mand match set
-                    segments = domain.split('.')
-                    if len(segments) < 2:  # what nonsense is this?
-                        continue
-
-                    # decompose 'zot.foo.bar.baz' into [ 'zot.foo.bar.baz', 'foo.bar.baz', 'bar.baz' ]
-                    checklist = ['.'.join(segments[i:]) for i in range(len(segments) - 1)]
-
-                    # check set membership.  This can be cheaper than any() if we get a match partway through the list
-                    found = False
-                    for item in checklist:
-                        if item in self.allow_domains_right:
-                            found = True
-                            break
-                    if found:
-                        continue
-
-                    # if we got here, there was no match, exact or right-hand.
-                    self.converter.write_line(output, line)
-                self.converter.after_writing(output)
-
-
-def all_subclasses(cls):
-    for subclass in cls.__subclasses__():
-        yield from all_subclasses(subclass)
-        yield subclass
-
-
-converters = {clazz.get_name(): clazz() for clazz in all_subclasses(RpzConverter)}
-
-
 def converter_choice(choice: str):
     try:
         return converters[choice]
     except KeyError:
         raise AttributeError(f'Bad converter value {choice}, valid choices are {converters.keys()}')
+
+
+# Give the user meaningful feedback if the script is being run as root without -U, or if the script has been
+# run with -U, but setuid support is not available on this platform.
+def check_suid(username: str):
+    try:
+        uid = os.getuid()
+        if uid == 0:
+            if username:
+                try:
+                    user = pwd.getpwnam(username)
+                    os.setuid(user.pw_uid)
+                except KeyError:
+                    print(f'User named {username} not found!')
+                    exit(1)
+                except NameError:
+                    print(f'-U specified, but no getpwnam support available on this system.')
+                    exit(1)
+            else:
+                print('Please do not run this script as root!  (See also the -U argument.)')
+                exit(1)
+    except AttributeError:
+        if username:
+            print('-U specified, but setuid support is not available on this system.')
+            exit(1)
 
 
 if __name__ == '__main__':
@@ -327,26 +86,7 @@ if __name__ == '__main__':
 
     args = parser.parse_args()
 
-    try:
-        uid = os.getuid()
-        if uid == 0:
-            if args.U:
-                try:
-                    user = pwd.getpwnam(args.U)
-                    os.setuid(user.pw_uid)
-                except KeyError:
-                    print(f'User named {args.U} not found!')
-                    exit(1)
-                except NameError:
-                    print(f'-U specified, but no getpwnam support available on this system.')
-                    exit(1)
-            else:
-                print('Please do not run this script as root!  (See also the -U argument.)')
-                exit(1)
-    except AttributeError:
-        if args.U:
-            print('-U specified, but setuid support is not available on this system.')
-            exit(1)
+    check_suid(args.U)
 
     processor = RpzProcessor(converter_choice(args.c))
